UI + Earth Mage Animations
Implemented the UI and started coding it. I also started building the
architecture with how the abilities will work together with the UI. All
of these are happening in a new separate level. I also added 3 new
animations for the Earth Mage for the 3 basic abilities.. The variables
for them need to change to 2, 3 and 4 where 0 is idle (which was also
updated) and 1 is walking.