Map, camera script and prefab update
I updated the map model so it no longer has that jump effect. I updated
the camera script and added it to the Arena1 level properly so it works
well. I updated the model prefab for the earth mage. There's a bug with
the camera touch on rotation which I'll try to fix. If you take one of
your fingers off but keep the other while rotating the camera bugs out.
This is not a new bug, just one I didn't notice in the past.