Updated the map
I like what it looks like, I think it's around 90% there but to be
honnest I'm not sure what else to do to it to improve it at this point.
I already wasted a bunch of time today trying to figure this out. I'll
work on other things and maybe latter on I can think of new elements
that I can add.